requirements.txt
asgiref==3.5.2
cloudinary==1.30.0
Django==3.2.16
django-cloudinary-storage==0.3.0
django-filter==22.1
django-language-field==0.0.3
djangorestframework==3.14.0
Pillow==9.3.0
pytz==2022.6
sqlparse==0.4.3
